Modify server files for eventual testing
- Split off part of server.ts to => expressApp.ts
- Uncomment some parts of jest.config.js for potential coverage reports
  - will be done via third party apps
- Add to : docs/notes.md
- Comment a route out in : routes/user.ts
